Split LearningJourney into focused components and remove unnecessary utilities
- Split main component into 6 focused sub-components:
  * ProgressCard, LevelButton, ComplementSection, OperationSection,
    CurrentQuestion, CheckAnswerButton
- Reduce main component from 345 to 108 lines (~69% reduction)
- Remove unnecessary classNames.ts utility file (42 lines)
- Inline className logic directly in components where used
- Improve component reusability and maintainability
- All 22 tests continue to pass without modification
